ml/random_forest.py
from sklearn.ensemble import RandomForestRegressor
import numpy as np

class RandomForestMutationModel:
    """Stub for Random Forest-based mutation prediction."""
    def __init__(self, model_path=None):
        self.model_path = model_path
        self.model = None
        if model_path:
            self.load_model(model_path)
    def load_model(self, path):
        # TODO: Load a real Random Forest model
        self.model = None
    def predict(self, features):
        # Accepts a dict or array; use the length of the first feature
        if isinstance(features, dict):
            arr = next(iter(features.values()))
            n = len(arr)
        else:
            n = features.shape[0]
        return np.random.rand(n)
    def train(self, X, y):
        # TODO: Implement training logic
        pass 