requirements.txt
streamlit
biopython
py3Dmol
plotly
numpy
scikit-learn
xgboost
torch
requests
nglview 